+++ main/scripts/deprecated/info.bat
