+++ main/scripts/functions/orbit.py
@@ -1,0 +1,118 @@
+from core import Core
+
+
+class Orbit(Core):
+
+    def set_apo(self, apo_target, wait_til_peri=True, burn_time=30, turn_time=5):
+
+        if wait_til_peri:
+            while self.vessel.orbit.time_to_periapsis > burn_time*3:
+                self.conn.space_center.rails_warp_factor = 3
+            while self.vessel.orbit.time_to_periapsis > burn_time:
+                self.conn.space_center.rails_warp_factor = 2
+
+        self.conn.space_center.rails_warp_factor = 0
+
+        print("Orienting craft")
+        if self.apoapsis() < apo_target:
+            self.sas_prograde()
+        elif self.apoapsis() > apo_target:
+            self.sas_retrograde()
+
+        time.sleep(turn_time)
+
+        print("Performing burn")
+        while self.apoapsis() < 0.95* apo_target:
+            self.vessel.control.throttle = 1
+        while self.apoapsis() < 0.99* apo_target:
+            self.vessel.control.throttle = 0.1
+        while self.apoapsis() < 0.99999* apo_target:
+            self.vessel.control.throttle = 0.01
+
+        self.vessel.control.throttle = 0
+        apo_actual = self.apoapsis()
+        print("Burn complete. Target apo: {}  Actual apo: {}".format(apo_target, apo_actual))
+
+
+    def set_peri(self, peri_target, wait_til_apo=True, burn_time=30, turn_time=5):
+        ## todo check if burn should be prograde or retrograde
+
+        if wait_til_apo:
+            while self.vessel.orbit.time_to_apoapsis > burn_time*3:
+                self.conn.space_center.rails_warp_factor = 3
+            while self.vessel.orbit.time_to_apoapsis > burn_time:
+                self.conn.space_center.rails_warp_factor = 2
+
+        self.conn.space_center.rails_warp_factor = 0
+
+        print("Orienting craft")
+        if self.periapsis() < peri_target:
+            self.sas_prograde()
+        elif self.periapsis() > peri_target:
+            self.sas_retrograde()
+
+        time.sleep(turn_time)
+
+        print("Performing burn")
+        while self.periapsis() < 0.95* peri_target:
+            self.vessel.control.throttle = 1
+        while self.periapsis() < 0.99* peri_target:
+            self.vessel.control.throttle = 0.1
+        while self.periapsis() < 0.99999* peri_target:
+            self.vessel.control.throttle = 0.01
+
+        self.vessel.control.throttle = 0
+        peri_actual = self.periapsis()
+        print("Burn complete. Target peri: {}  Actual peri: {}".format(peri_target, peri_actual))
+
+    def circularise_to_apo(self):
+
+        self.vessel.auto_pilot.engage()
+
+        mu = self.vessel.orbit.body.gravitational_parameter
+        r = self.vessel.orbit.apoapsis
+        a1 = self.vessel.orbit.semi_major_axis
+        a2 = r
+        v1 = math.sqrt(mu * ((2. / r) - (1. / a1)))
+        v2 = math.sqrt(mu * ((2. / r) - (1. / a2)))
+        delta_v = v2 - v1
+        node = self.vessel.control.add_node(self.ut() + self.vessel.orbit.time_to_apoapsis, prograde=delta_v)
+
+        F = self.vessel.available_thrust
+        Isp = self.vessel.specific_impulse * 9.82
+        m0 = self.vessel.mass
+        m1 = m0 / math.exp(delta_v / Isp)
+        flow_rate = F / Isp
+        burn_time = (m0 - m1) / flow_rate
+
+        print("Burn time: ", burn_time)
+
+        # Orientate ship
+        print('Orientating ship for circularization burn')
+        self.vessel.auto_pilot.reference_frame = node.reference_frame
+        self.vessel.auto_pilot.target_direction = (0, 1, 0)
+        self.vessel.auto_pilot.wait()
+
+        # Wait until burn
+        print('Waiting until circularization burn')
+        burn_ut = self.ut() + self.vessel.orbit.time_to_apoapsis - (burn_time / 2.)
+        lead_time = 5
+        self.conn.space_center.warp_to(burn_ut - lead_time)
+
+        # Execute burn
+        print('Ready to execute burn')
+        time_to_apoapsis = self.conn.add_stream(getattr, self.vessel.orbit, 'time_to_apoapsis')
+        while time_to_apoapsis() - (burn_time / 2.) > 0:
+            pass
+        print('Executing burn')
+        self.vessel.control.throttle = 1.0
+        time.sleep(burn_time - 0.1)
+        print('Fine tuning')
+        self.vessel.control.throttle = 0.05
+        remaining_burn = self.conn.add_stream(node.remaining_burn_vector, node.reference_frame)
+
+        while remaining_burn()[1] > 0.1:
+            pass
+        self.vessel.control.throttle = 0.0
+        node.remove()
+        self.sas_prograde()
