+++ main/scripts/functions/transfer.py
@@ -1,0 +1,6 @@
+from core import Core
+
+
+class Transfer(Core):
+
+    pass
