--- main/scripts/launcher.py
@@ -1,7 +1,0 @@
-from krpc_FUNCTIONS import Core, Launcher, Orbit, Lander
-
-launcher = Launcher()
-
-launcher.launch()
-launcher.gravity_turn()
-launcher.circularise()
