--- main/scripts/suicide_burn.py
@@ -1,5 +1,0 @@
-from krpc_FUNCTIONS import Core, Launcher, Orbit, Lander
-
-core = Lander()
-
-core.suicide_burn()
