+++ main/scripts/template_script.py
@@ -6,10 +6,3 @@ from lander import Lander
 from launcher import Launcher
 from orbit import Orbit
 from transfer import Transfer
-
-
-print("Running execute node script")
-core = Core()
-
-core.execute_next_node()
-print("Exiting execute node script")
